+++ classes/event/course_module_viewed.php
@@ -53,7 +53,7 @@ class course_module_viewed extends \core\event\course_module_viewed {
     public function get_url() {
         return new \moodle_url("/mod/$this->objecttable/view.php", [
             'id'  => $this->contextinstanceid,
-            'tab' => $this->data['other']['tab']
+            'tab' => $this->data['other']['tab'],
         ]);
     }
 }