+++ classes/event/queue_entry_created.php
@@ -100,7 +100,7 @@ class queue_entry_created extends \core\event\base {
         return new \moodle_url("/mod/grouptool/view.php", [
             'id'      => $this->contextinstanceid,
             'tab'     => 'overview',
-            'groupid' => $this->data['other']['groupid']
+            'groupid' => $this->data['other']['groupid'],
         ]);
     }