+++ mod_form.php
@@ -20,7 +20,8 @@
  * @package   mod_grouptool
  * @author    Philipp Hager
  * @author    Hannes Laimer
- * @copyright 2014 Academic Moodle Cooperation {@link http://www.academic-moodle-cooperation.org}
+ * @author    Anne Kreppenhofer
+ * @copyright 2024 Academic Moodle Cooperation {@link http://www.academic-moodle-cooperation.org}
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
@@ -34,7 +35,7 @@ require_once($CFG->dirroot.'/mod/grouptool/definitions.php');
  *
  * @package   mod_grouptool
  * @author    Philipp Hager
- * @copyright 2014 Academic Moodle Cooperation {@link http://www.academic-moodle-cooperation.org}
+ * @copyright 2024 Academic Moodle Cooperation {@link http://www.academic-moodle-cooperation.org}
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 class mod_grouptool_mod_form extends moodleform_mod {
@@ -135,7 +136,7 @@ class mod_grouptool_mod_form extends moodleform_mod {
                 GROUPTOOL_SHOW_GROUPMEMBERS_AFTER_DUE     => get_string('showafterdue', 'grouptool'),
                 GROUPTOOL_SHOW_OWN_GROUPMEMBERS_AFTER_DUE => get_string('showownafterdue', 'grouptool'),
                 GROUPTOOL_SHOW_OWN_GROUPMEMBERS_AFTER_REG => get_string('showownafterreg', 'grouptool'),
-                GROUPTOOL_SHOW_GROUPMEMBERS               => get_string('yes')
+                GROUPTOOL_SHOW_GROUPMEMBERS               => get_string('yes'),
         ];
         $mform->addElement('select', 'show_members', get_string('show_members', 'grouptool'), $options);
         $showmembers = get_config('mod_grouptool', 'show_members');
@@ -321,7 +322,7 @@ class mod_grouptool_mod_form extends moodleform_mod {
 
         $options = [
                 GROUPTOOL_IGNORE => get_string('ignorechanges', 'grouptool'),
-                GROUPTOOL_FOLLOW => get_string('followchanges', 'grouptool')
+                GROUPTOOL_FOLLOW => get_string('followchanges', 'grouptool'),
         ];
 
         $mform->addElement('select', 'ifmemberadded', get_string('ifmemberadded', 'grouptool'),
@@ -346,7 +347,7 @@ class mod_grouptool_mod_form extends moodleform_mod {
 
         $options = [
                 GROUPTOOL_RECREATE_GROUP => get_string('recreate_group', 'grouptool'),
-                GROUPTOOL_DELETE_REF => get_string('delete_reference', 'grouptool')
+                GROUPTOOL_DELETE_REF => get_string('delete_reference', 'grouptool'),
         ];
         $mform->addElement('select', 'ifgroupdeleted', get_string('ifgroupdeleted', 'grouptool'),
                            $options);
@@ -554,15 +555,16 @@ class mod_grouptool_mod_form extends moodleform_mod {
     public function add_completion_rules() {
         $mform =& $this->_form;
 
-        $group = array();
-        $group[] =& $mform->createElement('checkbox', 'completionregisterenabled', '',
-                get_string('completionregister', 'grouptool'));
-        $group[] =& $mform->createElement('text', 'completionregister', '', array('size' => 3));
-        $mform->setType('completionregister', PARAM_INT);
-        $mform->addGroup($group, 'completionregistergroup', get_string('require_registration', 'grouptool'), array(' '), false);
-        $mform->addHelpButton('completionregistergroup', 'require_registration', 'grouptool');
-        $mform->disabledIf('completionregister', 'completionregisterenabled', 'notchecked');
-        return ['completionregistergroup'];
+        $suffix = $this->get_suffix();
+
+        $group = [];
+        $group[] =& $mform->createElement('checkbox', 'completionregisterenabled'.$suffix, '', get_string('completionregister', 'grouptool'));
+        $group[] =& $mform->createElement('text', 'completionregister'.$suffix, '', ['size' => 3]);
+        $mform->setType('completionregister'.$suffix, PARAM_INT);
+        $mform->addGroup($group, 'completionregistergroup'.$suffix, get_string('require_registration', 'grouptool'), [' '], false);
+        $mform->addHelpButton('completionregistergroup'.$suffix, 'require_registration', 'grouptool');
+        $mform->disabledIf('completionregister'.$suffix, 'completionregisterenabled'.$suffix, 'notchecked');
+        return ['completionregistergroup'.$suffix];
     }
     /**
      * Determines if completion is enabled for this module.
@@ -571,7 +573,8 @@ class mod_grouptool_mod_form extends moodleform_mod {
      * @return bool
      */
     public function completion_rule_enabled($data) {
-        return (!empty($data['completionregisterenabled']) && $data['completionregister'] != 0);
+        $suffix = $this->get_suffix();
+        return (!empty($data['completionregisterenabled'.$suffix]) && $data['completionregister'.$suffix] != 0);
     }
 
     /**
@@ -584,6 +587,7 @@ class mod_grouptool_mod_form extends moodleform_mod {
      */
     public function data_postprocessing($data) {
         parent::data_postprocessing($data);
+        $suffix = $this->get_suffix();
         // Turn off completion settings if the checkboxes aren't ticked.
         if (!empty($data->completionunlocked)) {
             $autocompletion = !empty($data->completion) && $data->completion == COMPLETION_TRACKING_AUTOMATIC;